+++ packages/motion/src/index.js
@@ -1,2 +1,1 @@
-export { removeWithMotion } from "./remove-with-motion.js"
 export { withMotion } from "./with-motion.js"